requirements.txt
crewai>=0.175.0
crewai-tools>=0.15.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.6,<1.0.0
chromadb>=1.0.20
openai>=1.50.0
python-dotenv>=1.0.0
tiktoken>=0.8.0
pydantic>=2.10.0
PyPDF2>=3.0.0
